+++ packages/js/server/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @trikhub/server
 
+## 0.9.0
+
+### Minor Changes
+
+- [#30](https://github.com/Molefas/trikhub/pull/30) [`d3fb07e`](https://github.com/Molefas/trikhub/commit/d3fb07efb35563b08df2e5c73ed08d7676d99ab1) Thanks [@Molefas](https://github.com/Molefas)! - Updated & Cleaned up js HTTP server
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @trikhub/manifest@0.9.0
+  - @trikhub/gateway@0.9.0
+  - @trikhub/linter@0.9.0
+
 ## 0.7.0
 
 ### Patch Changes